:art: refactor: FloorSelector 컴포넌트 분리
- FloorSelector 컴포넌트를 src/app/components/FloorSelector.tsx 파일로 이동합니다.
